moving legacy files, start refactor Explorer
