+++ scqubits/core/explorer.py
@@ -12,7 +12,7 @@
 import matplotlib.pyplot as plt
 import numpy as np
 
-import scqubits.core.sweep_generators as swp
+import scqubits.legacy.sweep_generators as swp
 import scqubits.utils.explorer_panels as panels
 import scqubits.utils.misc as utils
