+++ scqubits/core/spec_lookup.py
@@ -10,10 +10,8 @@
 ############################################################################
 
 import itertools
-import warnings
 import weakref
 
-from functools import wraps
 from typing import TYPE_CHECKING, Callable, List, Tuple, Union
 
 import numpy as np
