+++ scqubits/legacy/sweep_observables.py
